pallet_revive_fixtures

Module Callee

Module containing a contract’s types and functions.

contract Callee {
    uint64 public stored;
    function echo(uint64 _data) external pure returns (uint64 data) { <stmts> }
    function whoSender() external view returns (address) { <stmts> }
    function store(uint64 _data) external { <stmts> }
    function revert() public pure returns (uint64) { <stmts> }
    function invalid() public pure { <stmts> }
    function stop() public pure { <stmts> }
    function consumeAllReftime() external { <stmts> }
}

Structs

consumeAllReftimeCall

Function with signature consumeAllReftime() and selector 0xb351696d.

consumeAllReftimeReturn

Container type for the return parameters of the consumeAllReftime() function.

echoCall

Function with signature echo(uint64) and selector 0x70363622.

echoReturn

Container type for the return parameters of the echo(uint64) function.

invalidCall

Function with signature invalid() and selector 0xbde9e40e.

invalidReturn

Container type for the return parameters of the invalid() function.

revertCall

Function with signature revert() and selector 0x7da3c3ab.

revertReturn

Container type for the return parameters of the revert() function.

stopCall

Function with signature stop() and selector 0x07da68f5.

stopReturn

Container type for the return parameters of the stop() function.

storeCall

Function with signature store(uint64) and selector 0x1d9a3bdd.

storeReturn

Container type for the return parameters of the store(uint64) function.

storedCall

Function with signature stored() and selector 0xe582dd31.

storedReturn

Container type for the return parameters of the stored() function.

whoSenderCall

Function with signature whoSender() and selector 0xca3e58fe.

whoSenderReturn

Container type for the return parameters of the whoSender() function.

Enums

CalleeCalls

Container for all the Callee function calls.