pallet_revive_fixtures

Module Caller

Module containing a contract’s types and functions.

contract Caller {
    uint256 public data;
    enum CallType { Call, StaticCall, DelegateCall }
    function normal(address _callee, uint64 _value, bytes memory _data, uint64 _gas) external returns (bool success, bytes memory output) { <stmts> }
    function delegate(address _callee, bytes memory _data, uint64 _gas) external returns (bool success, bytes memory output) { <stmts> }
    function staticCall(address _callee, bytes memory _data, uint64 _gas) external view returns (bool success, bytes memory output) { <stmts> }
    function create(bytes memory initcode) external payable returns (address addr) { <stmts> }
    function createRevert() external returns (address addr) { <stmts> }
    function create2(bytes memory initcode, bytes32 salt) external payable returns (address addr) { <stmts> }
    function callPartialGas(address _callee, bytes memory _data, uint64 _gasDivisor, CallType _callType) external returns (bool success) { <stmts> }
}

Structs

callPartialGasCall

Function with signature callPartialGas(address,bytes,uint64,uint8) and selector 0xb219b63b.

callPartialGasReturn

Container type for the return parameters of the callPartialGas(address,bytes,uint64,uint8) function.

create2Call

Function with signature create2(bytes,bytes32) and selector 0x8c0b8db2.

create2Return

Container type for the return parameters of the create2(bytes,bytes32) function.

createCall

Function with signature create(bytes) and selector 0xcf5ba53f.

createReturn

Container type for the return parameters of the create(bytes) function.

createRevertCall

Function with signature createRevert() and selector 0x51c18ef9.

createRevertReturn

Container type for the return parameters of the createRevert() function.

dataCall

Function with signature data() and selector 0x73d4a13a.

dataReturn

Container type for the return parameters of the data() function.

delegateCall

Function with signature delegate(address,bytes,uint64) and selector 0x066b54f3.

delegateReturn

Container type for the return parameters of the delegate(address,bytes,uint64) function.

normalCall

Function with signature normal(address,uint64,bytes,uint64) and selector 0x00d1df6a.

normalReturn

Container type for the return parameters of the normal(address,uint64,bytes,uint64) function.

staticCallCall

Function with signature staticCall(address,bytes,uint64) and selector 0x527e9ee8.

staticCallReturn

Container type for the return parameters of the staticCall(address,bytes,uint64) function.

Enums

CallType

CallerCalls

Container for all the Caller function calls.