pallet_revive_fixtures

Module Host

Module containing a contract’s types and functions.

contract Host {
    function balance(address account) public view returns (uint64) { <stmts> }
    function extcodesizeOp(address account) public view returns (uint64) { <stmts> }
    function extcodehashOp(address account) public view returns (bytes32) { <stmts> }
    function blockhashOp(uint64 blockNumber) public view returns (bytes32) { <stmts> }
    function sloadOp(uint64 slot) public view returns (uint64) { <stmts> }
    function sstoreOp(uint64 slot, uint64 value) public { <stmts> }
    function logOps() public { <stmts> }
    function selfbalance() public view returns (uint64) { <stmts> }
}

Structs

balanceCall

Function with signature balance(address) and selector 0xe3d670d7.

balanceReturn

Container type for the return parameters of the balance(address) function.

blockhashOpCall

Function with signature blockhashOp(uint64) and selector 0xc1dca71a.

blockhashOpReturn

Container type for the return parameters of the blockhashOp(uint64) function.

extcodehashOpCall

Function with signature extcodehashOp(address) and selector 0xe42cf66b.

extcodehashOpReturn

Container type for the return parameters of the extcodehashOp(address) function.

extcodesizeOpCall

Function with signature extcodesizeOp(address) and selector 0x5421a28a.

extcodesizeOpReturn

Container type for the return parameters of the extcodesizeOp(address) function.

logOpsCall

Function with signature logOps() and selector 0xe68d826f.

logOpsReturn

Container type for the return parameters of the logOps() function.

selfbalanceCall

Function with signature selfbalance() and selector 0xbf8e7c1c.

selfbalanceReturn

Container type for the return parameters of the selfbalance() function.

sloadOpCall

Function with signature sloadOp(uint64) and selector 0xe0400e82.

sloadOpReturn

Container type for the return parameters of the sloadOp(uint64) function.

sstoreOpCall

Function with signature sstoreOp(uint64,uint64) and selector 0x64b46e1d.

sstoreOpReturn

Container type for the return parameters of the sstoreOp(uint64,uint64) function.

Enums

HostCalls

Container for all the Host function calls.